update tile generation logic
- add planet scale tiles unpack
- add border handling for tiles unpack allows to avoid overriding trimmed tiles on border
- add MID_ZOOM for planet tile generation
- add MAX_PARALLEL_PSQL and COPY_CONCURRENCY to cpus count for tiles generation
- back `transportation_name` for planet tiles generation
- add params run tiles generation separate for belarus and planet

+++ updater/update.sh
@@ -15,14 +15,18 @@ mkdir -p testdata
 docker-compose run --rm management bash updater/upd_prepare_styles.sh
 docker-compose run --rm management bash updater/upd_prepare_fonts.sh
 
-docker-compose run --rm management bash updater/upd_download_belarus_dump.sh
-docker-compose run --rm management bash updater/upd_update_postgis.sh
-bash updater/upd_generate_belarus_mbtiles.sh
-docker-compose run --rm management bash updater/upd_unpack_belarus_mbtiles.sh
+if [ $1 = "all" ] || [ $1 = "belarus" ]; then
+  docker-compose run --rm management bash updater/upd_download_belarus_dump.sh
+  docker-compose run --rm management bash updater/upd_update_postgis.sh
+  bash updater/upd_generate_belarus_mbtiles.sh
+  docker-compose run --rm management bash updater/upd_unpack_belarus_mbtiles.sh
+fi
 
-docker-compose run --rm management bash updater/upd_download_planet_dump.sh
-bash updater/upd_generate_planet_mbtiles.sh
-docker-compose run --rm management bash updater/upd_unpack_planet_mbtiles.sh
+if [ $1 = "all" ] || [ $1 = "planet" ]; then
+  docker-compose run --rm management bash updater/upd_download_planet_dump.sh
+  bash updater/upd_generate_planet_mbtiles.sh
+  docker-compose run --rm management bash updater/upd_unpack_planet_mbtiles.sh
+fi
 
 docker-compose run --rm management bash updater/upd_prepare_tile_data_info.sh
